producer/requirements.txt
alpha_vantage
pandas
python-dotenv
requests